vfio/pci: Break up vfio_pci_core_ioctl() into one function per ioctl
500 lines is a bit long for a single function, move the bodies of each if
into separate functions and leave behind a switch statement to dispatch
them. Reindent all the code to drop a level of indentation. Everything
returns 'int' and doesn't use a 64 bit positive value.

No functional change.

issue: 2585478

+++ include/linux/vfio_pci_core.h
@@ -173,8 +173,8 @@ static inline ssize_t vfio_pci_vga_rw(struct vfio_pci_core_device *vdev,
 }
 #endif
 
-extern long vfio_pci_ioeventfd(struct vfio_pci_core_device *vdev, loff_t offset,
-			       uint64_t data, int count, int fd);
+extern int vfio_pci_ioeventfd(struct vfio_pci_core_device *vdev, loff_t offset,
+			      uint64_t data, int count, int fd);
 
 extern int vfio_pci_init_perm_bits(void);
 extern void vfio_pci_uninit_perm_bits(void);
